[space] Add ability to cap on the delta value
Some tasks should not be deferred indefinitely in the future. Their
delta value should instead either stop being incremented or be reset
when a cap is reached.

An example use case is habits. They should not be deferred say 20 days
into the future because that is not a habit. I want to see a habit in my
review often, but not every day (especially if I keep ignoring it) so I
think it might be healthy to keep a reset cap on it (which means that
the delta value resets when it reaches the cap).