Split el-secretario-keybindings into other files
The rationale for putting all the keybindings in the same file was
because I assumed you needed general.el for putting custom which-key
texts for a keybinding. I didn't want to have an additional obligatory
dependency so I put it into a separate melpa package. However that
assumption was false (see code in the commit for how) which is why I
split them up into the corresponding files.