fix: mu4e dashboard doesn't pop up after running el-secretario
After running el-secretario with a mu4e source, a user running mu4e the
`SPC o m` binding that usually brings up the dashboard would only get a
small bar at the bottom.  This was because el-secretario-mu4e was
setting the `mu4e-split-view' variable when activated.

This patch resets that variable to the value it had before when
el-secretario switches to the next source.  It's tricky to know when the
user is actually done with el-secretario, but I think this is the best
point to do so.