fix: caching issue with next-item-hook
The issue was that if you start a new session with an org source, and then a new
session without having modified the buffers, next-item-hook would not get
called.

org-ql caches queries when the neither the buffer or the query has been changed.
That messes with `el-secretario-org--parse-headline`, or rather,
`el-secretario-org--parse-headline` doesn't get called when the query is cached.
Instead, org-ql returns back the same objects that
`el-secretario-org--parse-headline` returned, objects that have since been
updated to keep track of whether next-item-hook has been called.

This commit disables the caching per the recommended method: https://github.com/alphapapa/org-ql/issues/437